+++ src/main/java/com/wanhive/iot/dao/UserDao.java
@@ -25,8 +25,12 @@ import com.wanhive.iot.provider.DataSourceProvider;
 public class UserDao {
 	public static PagedList list(long limit, long offset, String order, String orderBy, int type, int status)
 			throws SQLException, NamingException {
-		if (limit > Constants.getMaxItemsInList()) {
+		if (limit < 0 || limit > Constants.getMaxItemsInList()) {
 			throw new IllegalArgumentException("Invalid limit");
+		}
+
+		if (offset < 0) {
+			throw new IllegalArgumentException("Invalid offset");
 		}
 
 		StringBuilder queryBuilder = new StringBuilder();
@@ -97,11 +101,16 @@ public class UserDao {
 
 	public static PagedList search(String keyword, long limit, long offset, String order, String orderBy, int type,
 			int status) throws SQLException, NamingException {
-		if (limit > Constants.getMaxItemsInList()) {
+		if (limit < 0 || limit > Constants.getMaxItemsInList()) {
 			throw new IllegalArgumentException("Invalid limit");
 		}
 
-		if (keyword == null || keyword.length() < Constants.getMinSearchKeywordLength()) {
+		if (offset < 0) {
+			throw new IllegalArgumentException("Invalid offset");
+		}
+
+		if (keyword == null || keyword.length() < Constants.getMinSearchKeywordLength()
+				|| keyword.length() > Constants.getMaxSearchKeywordLength()) {
 			throw new IllegalArgumentException("Invalid keyword");
 		}