+++ ChangeLog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Maximum search keyword length in the application's settings.
+
+### Fixed
+
+- Check limit and offset before database query.
+- POJO fields should be private.
+
 ## [0.4.0] - 2021-02-01
 
 ### Added